third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-srcset-in-iframe-expected.txt

PASS window.internals.isLoadingFromMemoryCache("resources/image-set-4x.png") is false
PASS window.internals.isLoadingFromMemoryCache("resources/blue_rect.jpg") is true
PASS successfullyParsed is true

TEST COMPLETE
